requirements.txt
streamlit
pdfplumber
python-docx
pandas
plotly
spacy==3.7.2
thinc==8.1.10
numpy==1.26.4
fpdf
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
